requirement.txt
flask
pyjwt
python-dotenv
requests 
Extended